Initial setup for phase 6

+++ shared/src/main/java/model/ListGamesResult.java
@@ -2,4 +2,4 @@ package model;
 
 import chess.ChessGame;
 
-public record ListGamesResult(int gameID, String whiteUsername, String blackUsername, String gameName) {}
+public record ListGamesResult(double gameID, String whiteUsername, String blackUsername, String gameName) {}
